Output Syntax Error location to console (#26)
* fix: print error location in console

* chore: add keywords in package.json

* Add test cases for syntax-error
