+++ packages/dbml-cli/__test__/dbml2sql/syntax-error/in-files/ecommerce.dbml
@@ -1,0 +1,142 @@
+Table staff {
+  id int [pk]
+  first_name varchar
+  last_name varchar
+  address_id int
+  picture blob
+  email varchar
+  store_id int [ref: > store.id]
+  active boolean
+  user_name varchar
+  password varchar
+  last_update timestamp
+}
+
+Table store {
+  id int [pk
+  manager_staff_id int [ref: > staff.id]
+  address_id int
+  last_update timestamp
+}
+
+Table payment {
+  id int [pk]
+  customer_id int
+  staff_id int [ref: > staff.id]
+  rental_id int [ref: > rental.id]
+  amount decimal
+  payment_date datetime
+  last_update timestamp
+}
+
+Table rental {
+  id int [pk]
+  rental_date datetime
+  inventory_id int
+  customer_id int
+  return_date ddatetime
+  staff_id int [ref: > staff.id]
+  last_update timestamp
+}
+
+Table country {
+  id int [pk]
+  country varchar
+  last_update timestamp
+}
+
+Table city {
+  id int [pk]
+  city varchar
+  country_id int [ref: > country.id]
+  last_update timestamp
+}
+
+Table address {
+  id int [pk]
+  address varchar
+  address2 varchar
+  district varchar
+  city_id int [ref: > city.id]
+  postal_code varchar
+  phone varchar
+  last_update timestamp
+}
+
+Table customer {
+  id int [pk]
+  store_id int
+  first_name varchar
+  last_name varchar
+  email varchar
+  address_id int [ref: > address.id]
+  active boolean
+  create_Date timestamp
+  last_update timestamp
+
+  Indexes {
+    (id, first_name) [type: btree]
+  }
+}
+
+Table category {
+  id int [pk]
+  name varchar
+  last_update timestamp
+}
+
+Table film_category {
+  id int [pk]
+  category_id int [ref: > category.id]
+  last_update timestamp
+}
+
+Table language {
+  id int [pk]
+  name varchar
+  last_update timestamp
+}
+
+Table film_text {
+  id int [pk]
+  film_id int [ref: > inventory.film_id]
+  title varchar
+  description text
+}
+
+Table actor {
+  id int [pk]
+  first_name varchar
+  last_name varchar
+  last_update timestamp
+}
+
+Table film {
+  id int [pk]
+  title varchar
+  description text
+  releaase_year int
+  language_id int [ref: > language.id]
+  original_language_id int [ref: > language.id]
+  rental_duration int 
+  rental_rate decimal
+  length int
+  replacement_cost decimal
+  rating varchar
+  special_feature varchar
+  last_update timestamp
+}
+
+Table film_actor {
+  id int [pk]
+  film_id int [ref: > film.id]
+  actor_id int [ref: > actor.id]
+  last_update timestamp
+}
+
+Table inventory {
+  id int [pk]
+  film_id int [ref: > film.id]
+  store_id int 
+  last_update timestamp
+}
