+++ packages/dbml-cli/__test__/dbml2sql/syntax-error/options.json
@@ -1,0 +1,5 @@
+{
+  "args": [
+    "./in-files/ecommerce.dbml"
+  ]
+}
