+++ packages/dbml-cli/__test__/dbml2sql/syntax-error/stdout.txt
@@ -1,0 +1,3 @@
+  ERROR: You have a syntax error at "ecommerce.dbml" line 17 column 3. Expected ",", "]", comment, or whitespace but "m" found.
+
+  A complete log can be found in:
