+++ packages/dbml-cli/__test__/sql2dbml/syntax-error/options.json
@@ -1,0 +1,6 @@
+{
+  "args": [
+    "--mysql",
+    "./in-files/business.sql"
+  ]
+}
