+++ packages/dbml-cli/__test__/sql2dbml/syntax-error/stdout.txt
@@ -1,0 +1,3 @@
+  ERROR: You have a syntax error at "business.sql" line 5 column 3. Expected ")" or fields but "`" found.
+
+  A complete log can be found in:
