+++ packages/dbml-cli/src/errors/domainError.js
@@ -1,0 +1,10 @@
+class DomainError extends Error {
+  constructor (message, rootError = {}) {
+    super(message);
+    this.name = this.constructor.name;
+    this.rootError = rootError;
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+export default DomainError;
