+++ packages/dbml-core/package.json
@@ -1,11 +1,15 @@
 {
   "name": "@dbml/core",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "description": "> TODO: description",
   "author": "Holistics <dev@holistics.io>",
   "license": "Apache-2.0",
   "homepage": "https://dbml-lang.org/home/",
   "repository": "https://github.com/holistics/dbml/tree/master/packages/dbml-core",
+  "keywords": [
+    "dbml",
+    "dbml-core"
+  ],
   "directories": {
     "lib": "lib",
     "test": "__tests__"
